SDI-2324-804-session10-Ejercicio-complementario1-ACABADO

+++ musicstoreapp/routes/api/songsAPIv1.0.js
@@ -40,18 +40,30 @@ module.exports = function (app, songsRepository, usersRepository) {
         try {
             let songId = new ObjectId(req.params.id)
             let filter = {_id: songId}
-            songsRepository.deleteSong(filter, {}).then(result => {
-                if (result === null || result.deletedCount === 0) {
-                    res.status(404);
-                    res.json({error: "ID inválido o no existe, no se ha borrado el registro."});
-                } else {
-                    res.status(200);
-                    res.send(JSON.stringify(result));
+
+            //  asegurarse de que el usuario es el dueño de la canción
+            //  asegurarse de que el usuario es el dueño de la canción
+            userIsAuthor(res.user, songId).then(isAuthor => {
+                if (isAuthor) {
+                    songsRepository.deleteSong(filter, {}).then(result => {
+                        if (result === null || result.deletedCount === 0) {
+                            res.status(404);
+                            res.json({error: "ID inválido o no existe, no se ha borrado el registro."});
+                        } else {
+                            res.status(200);
+                            res.send(JSON.stringify(result));
+                        }
+                    }).catch(error => {
+                        res.status(500);
+                        res.json({error: "Se ha producido un error al eliminar la canción."})
+                    });
                 }
-            }).catch(error => {
-                res.status(500);
-                res.json({error: "Se ha producido un error al eliminar la canción."})
-            });
+                else {
+                    res.status(500);
+                    res.json({errors: "La canción que se intenta eliminar no es del usuario autenticado."})
+                }
+            })
+
         } catch (e) {
             res.status(500);
             res.json({error: "Se ha producido un error, revise que el ID sea válido."})
@@ -64,7 +76,7 @@ module.exports = function (app, songsRepository, usersRepository) {
                 title: req.body.title,
                 kind: req.body.kind,
                 price: req.body.price,
-                author: req.session.user
+                author: res.user
             }
             // Validar aquí: título, género, precio y autor.
             validatorInsertSong(song, function (errors) {
@@ -100,18 +112,8 @@ module.exports = function (app, songsRepository, usersRepository) {
             //Si la _id NO no existe, no crea un nuevo documento.
             const options = {upsert: false};
             let song = {
-                author: req.session.user
+                author: res.user
             }
-
-
-            validatorEditSong(song, function (errors) {
-                if (errors !== null && errors.length > 0) {
-                    res.status(422);
-                    res.json({errors: errors})
-                } else {
-
-                }
-            })
 
             if (typeof req.body.title !== "undefined" && req.body.title !== null)
                 song.title = req.body.title;
@@ -119,27 +121,47 @@ module.exports = function (app, songsRepository, usersRepository) {
                 song.kind = req.body.kind;
             if (typeof req.body.price !== "undefined" && req.body.price !== null)
                 song.price = req.body.price;
-            songsRepository.updateSong(song, filter, options).then(result => {
-                if (result === null) {
-                    res.status(404);
-                    res.json({error: "ID inválido o no existe, no se ha actualizado la canción."});
-                }
-                //La _id No existe o los datos enviados no difieren de los ya almacenados.
-                else if (result.modifiedCount == 0) {
-                    res.status(409);
-                    res.json({error: "No se ha modificado ninguna canción."});
-                }
-                else{
-                    res.status(200);
-                    res.json({
-                        message: "Canción modificada correctamente.",
-                        result: result
+
+            //  asegurarse de que el usuario es el dueño de la canción
+            userIsAuthor(res.user, songId).then(isAuthor => {
+                if (isAuthor) {
+
+                    // Validar aquí: título, género, precio y autor.
+                    validatorInsertSong(song, function (errors) {
+                        if (errors !== null && errors.length > 0) {
+                            res.status(422);
+                            res.json({errors: errors})
+                        } else {
+                            songsRepository.updateSong(song, filter, options).then(result => {
+                                if (result === null) {
+                                    res.status(404);
+                                    res.json({error: "ID inválido o no existe, no se ha actualizado la canción."});
+                                }
+                                //La _id No existe o los datos enviados no difieren de los ya almacenados.
+                                else if (result.modifiedCount == 0) {
+                                    res.status(409);
+                                    res.json({error: "No se ha modificado ninguna canción."});
+                                } else {
+                                    res.status(200);
+                                    res.json({
+                                        message: "Canción modificada correctamente.",
+                                        result: result
+                                    })
+                                }
+                            }).catch(error => {
+                                res.status(500);
+                                res.json({error: "Se ha producido un error al modificar la canción."})
+                            });
+                        }
                     })
                 }
-            }).catch(error => {
-                res.status(500);
-                res.json({error : "Se ha producido un error al modificar la canción."})
+                else {
+                    res.status(500);
+                    res.json({errors: "La canción que se intenta eliminar no es del usuario autenticado."})
+                }
             });
+
+
         } catch (e) {
             res.status(500);
             res.json({error: "Se ha producido un error al intentar modificar la canción: "+ e})
@@ -225,35 +247,17 @@ module.exports = function (app, songsRepository, usersRepository) {
     }
 
 
-        function validatorUpdateSong(song, callbackFunction) {
-            let errors = new Array();
-            if (song.title === null || typeof song.title === 'undefined' || song.title.trim().length == 0)
-                errors.push({
-                    "value": song.title,
-                    "msg": "El titulo de la canción no puede estar vacio",
-                    "param": "title",
-                    "location": "body"
-                })
-            if (song.kind === null || typeof song.kind === 'undefined' || song.kind.trim().length == 0)
-                errors.push({
-                    "value": song.kind,
-                    "msg": "el género de la canción no puede estar vacio",
-                    "param": "kind",
-                    "location": "body"
-                })
-            if (song.price === null || typeof song.price === 'undefined' || song.price < 0 ||
-                song.price.toString().trim().length == 0)
-                errors.push({
-                    "value": song.price,
-                    "msg": "El precio de la canción no puede ser negativo",
-                    "param": "price",
-                    "location": "body"
-                })
-            if (errors.length <= 0)
-                callbackFunction(null);
-            else
-                callbackFunction(errors);
-        }
+    async function userIsAuthor(user, songId) {
+
+        let filter = {_id: songId};
+        let foundSong = await songsRepository.findSong(filter, {})
+        console.log(foundSong);
+        console.log(user);
+        if (foundSong == null || foundSong.author !== user || foundSong.length > 0 )
+            return false;
+        return true;
+
+    }
 
 }
